ゲームスピードに合わせたカウントの進み方にしました

+++ Assets/NT/NT_Script/NT_graph.cs
@@ -5,7 +5,7 @@ namespace NT
 {
     public class NT_graph : MiniGameBase
     {
-        private int count;
+        private float count;
         [SerializeField] private Image image;
         [SerializeField] private NT_switch switchComponent;
